refactor : Consumer 파일 분리
consumers 폴더 추가
game, waiting 부분으로 분리
